feat: better remaster distinction, showing (#30)

+++ migrations/20250119092802_fuzzy_and_encounters.sql
@@ -19,7 +19,7 @@ CREATE TABLE campaign_sessions (
     -- Non-metadata or unique. A collection of unassigned rewards in linked encounters.
     -- eg: The gold in a session is the sum of all the gold in the encounters, minus the gold assigned to characters.
     unassigned_gold_rewards double precision NOT NULL DEFAULT 0,
-    unassigned_item_rewards bigint[] NOT NULL DEFAULT '{}'
+    unassigned_item_rewards integer[] NOT NULL DEFAULT '{}'
 );
 
 ALTER TABLE encounters ADD COLUMN session_id INTEGER NULL REFERENCES campaign_sessions(id);
@@ -32,7 +32,7 @@ CREATE TABLE campaign_session_characters (
     character_id INTEGER NOT NULL REFERENCES characters(id),
 
     gold_rewards double precision NOT NULL DEFAULT 0,
-    item_rewards bigint[] NOT NULL DEFAULT '{}',
+    item_rewards integer[] NOT NULL DEFAULT '{}',
 
     PRIMARY KEY (session_id, character_id)
 );
